requirements.txt
streamlit
twelvelabs
requests
python-dotenv
moviepy